Expose hoursToday function to $scope
Rather that set an hoursToday property on the instance of a location,
expose the underlyinf utility function to the scope so we can use it to
format the hours for a research location as well as it's divisions.

This needs to be carried through to the other places that display hours,
or maybe made global.