Split up askdonatefooter directive
Rather that have one directive to create the whole page section with the
donate, chat, and email buttons, split this up into separate directives
for each element so that we can move the individual items around or
reuse them as necessary.

It will probably be convenient to create a superdirective to create that
section uniformly, once the dust has settled.