Fix repeated hours-today ids
For a location with multiple divisions, each division has its own hours and,
since the todaysHours directive is used to generate them, they all have the same
id. The id needed on the main hours box because the id is used as a CSS
selector.

Moves the id onto the invocation of the directive for the main hours box and
removes it from the directive template. Therefore the hours for each division
have no ids (they're unneccesary anyway) and the id is still there to be used as
a selector for the main location hours.